refactor: Update import paths for organization-related entities
