+++ apps/klubiq-dashboard/src/lease/repositories/lease.repository.ts
@@ -15,7 +15,7 @@ import { CreateTenantDto } from '@app/common/dto/requests/create-tenant.dto';
 import { TenantUser } from '@app/common/database/entities/tenant.entity';
 import ShortUniqueId from 'short-unique-id';
 import { PropertyLeaseMetrics } from '../dto/responses/view-lease.dto';
-import { Unit } from '../../properties/entities/unit.entity';
+import { Unit } from '@app/common/database/entities/unit.entity';
 import { LeaseStatus } from '@app/common';
 
 @Injectable()
