+++ apps/klubiq-dashboard/src/organization/organization.module.ts
@@ -1,8 +1,6 @@
 import { Module } from '@nestjs/common';
-import { TypeOrmModule } from '@nestjs/typeorm';
 import { OrganizationService } from './services/organization.service';
 import { OrganizationController } from './controllers/organization.controller';
-import { Organization } from './entities/organization.entity';
 import { OrganizationRepository } from './repositories/organization.repository';
 import { EntityManager } from 'typeorm';
 import { OrganizationProfile } from './profiles/organization-profile';
@@ -18,7 +16,6 @@ import { OrganizationProfile } from './profiles/organization-profile';
 			inject: [EntityManager],
 		},
 	],
-	imports: [TypeOrmModule.forFeature([Organization])],
 	exports: [OrganizationRepository, OrganizationService],
 })
 export class OrganizationModule {}
