+++ apps/klubiq-dashboard/src/organization/profiles/organization-profile.ts
@@ -1,7 +1,7 @@
 import { UpdateOrganizationDto } from '../dto/requests/update-organization.dto';
 import { Mapper, MappingProfile, createMap } from '@automapper/core';
 import { AutomapperProfile, InjectMapper } from '@automapper/nestjs';
-import { Organization } from '../entities/organization.entity';
+import { Organization } from '@app/common/database/entities/organization.entity';
 import { CreateOrganizationDto } from '../dto/requests/create-organization.dto';
 
 export class OrganizationProfile extends AutomapperProfile {
