+++ apps/klubiq-dashboard/src/organization/repositories/organization.repository.ts
@@ -1,4 +1,4 @@
-import { Organization } from '../entities/organization.entity';
+import { Organization } from '@app/common/database/entities/organization.entity';
 import {
 	Injectable,
 	Logger,
@@ -8,7 +8,7 @@ import {
 import { BaseRepository, UserRoles } from '@app/common';
 import { EntityManager } from 'typeorm';
 import { UpdateOrganizationDto } from '../dto/requests/update-organization.dto';
-import { OrganizationUser } from '../../users/entities/organization-user.entity';
+import { OrganizationUser } from '@app/common/database/entities/organization-user.entity';
 
 @Injectable()
 export class OrganizationRepository extends BaseRepository<Organization> {
