+++ apps/klubiq-dashboard/src/users/controllers/users.controller.ts
@@ -10,7 +10,7 @@ import {
 import { UsersService } from '../services/users.service';
 
 import { ApiBearerAuth, ApiOkResponse, ApiTags } from '@nestjs/swagger';
-import { OrganizationUser } from '../entities/organization-user.entity';
+import { OrganizationUser } from '@app/common/database/entities/organization-user.entity';
 import { Actions, AppFeature, UserRoles } from '@app/common';
 import { AuthType, Auth, Roles, Feature, Ability } from '@app/auth';
 import { UpdateUserDto } from '../dto/update-organization-user.dto';
