+++ libs/auth/src/services/landlord-auth.service.ts
@@ -33,8 +33,8 @@ import {
 import { UserProfilesRepository } from '@app/common/repositories/user-profiles.repository';
 import { EmailTemplates } from '@app/common/email/types/email.types';
 import { ErrorMessages } from '@app/common/config/error.constant';
-import { Organization } from '../../../../apps/klubiq-dashboard/src/organization/entities/organization.entity';
-import { OrganizationUser } from '../../../../apps/klubiq-dashboard/src/users/entities/organization-user.entity';
+import { Organization } from '@app/common/database/entities/organization.entity';
+import { OrganizationUser } from '@app/common/database/entities/organization-user.entity';
 import { Mapper } from '@automapper/core';
 import { FirebaseErrorMessageHelper } from '../helpers/firebase-error-helper';
 import { ConfigService } from '@nestjs/config';
