+++ libs/common/src/database/entities/property-status.entity.ts
@@ -1,7 +1,7 @@
 import { Entity, Column, OneToMany } from 'typeorm';
 import { AutoMap } from '@automapper/classes';
 import { AbstractEntity } from './abstract-entity';
-import { Property } from '../../../../../apps/klubiq-dashboard/src/properties/entities/property.entity';
+import { Property } from './property.entity';
 
 @Entity({ schema: 'kdo' })
 export class PropertyStatus extends AbstractEntity {
